requirements.txt
azure-cognitiveservices-speech>=1.31.0
openai>=1.3.7
Flask>=3.0.0
